feat(ui): riepilogo terminal in navbar, data mese esteso, scala tipografica, sfondo piu' deciso
- navbar: striscia centrale di riepilogo (in attesa / assegnate / partite +
  banchine occupate/totali) via GET /api/system/summary; aggiorna a ogni Next
  Day e ogni 12s. Riempie il 'centro vuoto' con info utili a tutti i ruoli
- data: mese per intero ('20 giugno', non piu' 'giu'); giorno come 'Giorno N'
  in navbar (niente maiuscolo/zero-pad), 'gN' nelle tabelle, 'D/M' nella timeline
- scala tipografica in tokens (--fs-xs..--fs-xl), applicata ai titoli h2/h3
- sfondo mare piu' deciso: gradiente piu' profondo, onde su 3 livelli piu' marcate

+++ blueharbor_sprint2/frontend/src/services/time.js
@@ -21,10 +21,11 @@ export function formatDay(dayNumber, { mode = 'day', day1Date = null, withPrefix
   if (mode === 'date') {
     const d = dayToDate(dayNumber, day1Date);
     if (d) {
-      // compact = numerico "20/6" (per le colonne strette della timeline).
+      // compact = numerico "20/6" (per le colonne strette della timeline);
+      // esteso = "20 giugno" (mese per intero, non abbreviato).
       return compact
         ? `${d.getDate()}/${d.getMonth() + 1}`
-        : d.toLocaleDateString('it-IT', { day: 'numeric', month: 'short' });
+        : d.toLocaleDateString('it-IT', { day: 'numeric', month: 'long' });
     }
   }
   return withPrefix ? `g${dayNumber}` : String(dayNumber);
